refactor: split ui profile manager
